+++ include/simgrid/host.h
@@ -91,8 +91,8 @@ XBT_PUBLIC const char* sg_host_self_get_name();
  */
 XBT_PUBLIC unsigned long sg_host_get_nb_pstates(const_sg_host_t host);
 
-XBT_PUBLIC int sg_host_get_pstate(const_sg_host_t host);
-XBT_PUBLIC void sg_host_set_pstate(sg_host_t host, int pstate);
+XBT_PUBLIC unsigned long sg_host_get_pstate(const_sg_host_t host);
+XBT_PUBLIC void sg_host_set_pstate(sg_host_t host, unsigned long pstate);
 
 XBT_PUBLIC void sg_host_turn_on(sg_host_t host);
 XBT_PUBLIC void sg_host_turn_off(sg_host_t host);
